requirements.txt
flask
requests
azure-ai-documentintelligence